:recycle: fix minor issues in header file

+++ include/shell/helpers.h
@@ -9,7 +9,7 @@
     #define HELPERS_H_
 
 long file_get_size(char const *filepath);
-char *file_read_fd(int fd, long filesize);
+char *file_read_fd(int fd, size_t filesize);
 char *file_read(char *filepath);
 
 #endif /* !HELPERS_H_ */
